fix deps
