+++ sentryflow-clients/log-client/go.mod
@@ -2,16 +2,15 @@ module log-client
 
 go 1.19
 
-replace sentryflow/protobuf => ../protobuf
+replace sentryflow/protobuf => ../../protobuf
 
 require (
-	google.golang.org/grpc v1.61.0
+	google.golang.org/grpc v1.61.1
 	sentryflow/protobuf v0.0.0-00010101000000-000000000000
 )
 
 require (
 	github.com/golang/protobuf v1.5.3 // indirect
-	github.com/google/go-cmp v0.6.0 // indirect
 	golang.org/x/net v0.20.0 // indirect
 	golang.org/x/sys v0.16.0 // indirect
 	golang.org/x/text v0.14.0 // indirect
