+++ sentryflow-clients/mongo-client/go.mod
@@ -2,11 +2,11 @@ module mongo-client
 
 go 1.19
 
-replace sentryflow/protobuf => ../protobuf
+replace sentryflow/protobuf => ../../protobuf
 
 require (
 	go.mongodb.org/mongo-driver v1.13.1
-	google.golang.org/grpc v1.61.0
+	google.golang.org/grpc v1.61.1
 	sentryflow/protobuf v0.0.0-00010101000000-000000000000
 )
